refactor(cli): move helpers into _helpers/ subpackage (#168)
Separate CLI helper modules from command modules by moving them under
``_helpers/``. The underscore signals internal-only — external code
should not depend on these paths.

- ``cli/notification_setup.py`` → ``cli/_helpers/notification_setup.py``
- ``cli/transport_options.py``  → ``cli/_helpers/transport_options.py``
- ``cli/submission_plan.py`` (Phase 8f back-compat shim) deleted —
  callers routed to the canonical ``srunx.runtime.submission_plan``
  (only the two test files that used the shim path were updated).

All monkey-patch targets (``patch('srunx.cli.notification_setup.*')``)
updated in ``tests/test_callbacks.py`` + ``tests/transport/test_review_fixes.py``.
Docstring cross-references in ``callbacks.py`` and ``sync/hash_verify.py``
updated accordingly.

Closes #168
