+++ src/srunx/cli/_helpers/notification_setup.py
