+++ src/srunx/cli/commands/reports.py
@@ -7,7 +7,7 @@ import typer
 from rich.console import Console
 from rich.table import Table
 
-from srunx.cli.transport_options import LocalOpt, ProfileOpt, QuietOpt
+from srunx.cli._helpers.transport_options import LocalOpt, ProfileOpt, QuietOpt
 from srunx.logging import get_logger
 
 logger = get_logger(__name__)
